refactor(document.js): rename document model file
